And now, JOS 64

+++ inc/types.h
@@ -24,8 +24,8 @@ typedef unsigned long long uint64_t;
 // uintptr_t to represent the numerical values of virtual addresses,
 // and physaddr_t to represent physical addresses.
 typedef int32_t intptr_t;
-typedef uint32_t uintptr_t;
-typedef uint32_t physaddr_t;
+typedef uint64_t uintptr_t;
+typedef uint64_t physaddr_t;
 
 // Page numbers are 32 bits long.
 typedef uint32_t ppn_t;
@@ -57,14 +57,14 @@ typedef int32_t off_t;
 // Round down to the nearest multiple of n
 #define ROUNDDOWN(a, n)						\
 ({								\
-	uint32_t __a = (uint32_t) (a);				\
+	uint64_t __a = (uint64_t) (a);				\
 	(typeof(a)) (__a - __a % (n));				\
 })
 // Round up to the nearest multiple of n
 #define ROUNDUP(a, n)						\
 ({								\
-	uint32_t __n = (uint32_t) (n);				\
-	(typeof(a)) (ROUNDDOWN((uint32_t) (a) + __n - 1, __n));	\
+	uint64_t __n = (uint64_t) (n);				\
+	(typeof(a)) (ROUNDDOWN((uint64_t) (a) + __n - 1, __n));	\
 })
 
 // Return the offset of 'member' relative to the beginning of a struct type
